pyproject.toml
[project]
name = "skillsarathi-ai"
version = "0.1.0"
description = "Skillsarathi AI - Advanced Multi-Agent AI Companion Platform for India"
authors = [
    {name = "Skillsarathi Team", email = "team@skillsarathi.ai"},
]
readme = "README.md"
dependencies = [
    "fastapi>=0.116.0",
    "uvicorn[standard]>=0.35.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg2-binary>=2.9.9",
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "pydantic[email]>=2.11.0",
    "pydantic-settings>=2.10.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "passlib[bcrypt]>=1.7.4",
    "langchain>=0.3.0",
    "langchain-community>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.2.0",
    "chromadb>=1.0.20",
    "sentence-transformers>=5.1.0",
    "pypdf>=3.17.4",
    "python-docx>=1.1.0",
    "aiohttp>=3.12.0",
    "redis>=5.0.1",
    "celery>=5.3.4",
    "httpx>=0.28.0",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    "python-dotenv>=1.1.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "email-validator>=2.2.0",
    "starlette>=0.47.0",
    "bcrypt>=4.3.0",
    "langchain-core>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "tenacity>=9.1.0",
    "tqdm>=4.67.0",
    "openai>=1.100.0",
    "anthropic>=0.7.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
    "tiktoken>=0.7.0",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "typing-extensions>=4.8.0",
    "packaging>=23.2",
    "streamlit>=1.48.1",
    "websockets>=15.0.1",
    "streamlit-webrtc>=0.63.4",
    "av>=14.4.0",
    "opencv-python>=4.11.0.86",
    "speechrecognition>=3.14.3",
    "pyaudio>=0.2.14",
]
requires-python = ">= 3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
